+++ src/index.ts
@@ -13,6 +13,7 @@ import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
 
 import { VERSION } from "./api";
+import { registerAnalytics } from "./commands/analytics";
 import { registerAuth } from "./commands/auth";
 import { registerComments } from "./commands/comments";
 import { registerCrud } from "./commands/crud";
@@ -64,5 +65,6 @@ cli = registerMedia(cli);
 cli = registerPosts(cli);
 cli = registerComments(cli);
 cli = registerInbox(cli);
+cli = registerAnalytics(cli);
 
 cli.parse();